Do not remove localhost for unsuccessful local JMX connection
git-svn-id: https://svn.java.net/svn/visualvm~svn/trunk@2086 ba0df5e5-9053-6d9f-941f-f112716aa91a

+++ visualvm/jmx/src/com/sun/tools/visualvm/jmx/impl/JmxApplicationProvider.java
@@ -299,7 +299,7 @@ public class JmxApplicationProvider {
         // NOTE: this is not absolutely failsafe, if resolving the JMX application
         // took a long time and its host has been added by the user/plugin, it may
         // be removed by this call. Hopefully just a hypothetical case...
-        if (host != null && !hosts.contains(host))
+        if (host != null && !Host.LOCALHOST.equals(host) && !hosts.contains(host))
             host.getOwner().getRepository().removeDataSource(host);
     }
